Add computational methods and remove concrete engines.
This commit adds a number of computational methods to the metadata.
It removes previously defined engine names, pre/post processors and
visualizer. It replaces them with a generic engine metadata structure
and a set of engine features which are supposed to be filled by each
wrapper separately, as a set of key/value representing model/methods.